[FEAT] Add email input, email confirm & welcome page. Fix autologin feature.
