--- lib/screens/login/login/login_pw_input.dart
@@ -1,48 +1,0 @@
-import 'package:eva_icons_flutter/eva_icons_flutter.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-import '../../../config/palette.dart';
-
-class LoginPWInput extends StatelessWidget {
-  final String email;
-
-  const LoginPWInput({Key? key, required this.email}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      resizeToAvoidBottomInset: false,
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        leadingWidth: 80,
-        leading: IconButton(
-          onPressed: () {
-            Navigator.of(context).pop();
-          },
-          icon: const Icon(
-            EvaIcons.arrowIosBack,
-            color: Palette.black,
-          ),
-          iconSize: 40,
-        ),
-        backgroundColor: Colors.white,
-        systemOverlayStyle: SystemUiOverlayStyle.dark,
-        elevation: 0,
-        toolbarHeight: 100,
-      ),
-      body: GestureDetector(
-        onTap: () {
-          FocusScope.of(context).requestFocus(FocusNode());
-        },
-        child: Container(
-          height: double.infinity,
-          color: Colors.white,
-          child: const Center(
-            child: Text("로그인 패스워드"),
-          ),
-        ),
-      ),
-    );
-  }
-}
