+++ lib/utils/transition.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+void navigateHorizontally({required BuildContext context, required Widget widget})  {
+  Navigator.push(
+    context,
+    PageRouteBuilder(
+      pageBuilder:
+          (context, animation, secondaryAnimation) =>
+          widget,
+      transitionsBuilder: (context, animation,
+          secondaryAnimation, child) {
+        var begin = const Offset(1.0, 0.0);
+        var end = Offset.zero;
+        var curve = Curves.ease;
+
+        var tween = Tween(begin: begin, end: end)
+            .chain(CurveTween(curve: curve));
+
+        return SlideTransition(
+          position: animation.drive(tween),
+          child: child,
+        );
+      },
+    ),
+  );
+}
