+++ lib/widgets/next_button.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:wooyeon_flutter/config/palette.dart';
+import 'package:wooyeon_flutter/utils/transition.dart';
 
 class NextButton extends StatelessWidget {
   final Widget? nextPage;
@@ -23,27 +24,7 @@ class NextButton extends StatelessWidget {
             onTap: () {
               if (isActive) {
                 if (func == null && nextPage != null) {
-                  Navigator.push(
-                    context,
-                    PageRouteBuilder(
-                      pageBuilder: (context, animation, secondaryAnimation) =>
-                      nextPage!,
-                      transitionsBuilder:
-                          (context, animation, secondaryAnimation, child) {
-                        var begin = const Offset(1.0, 0.0);
-                        var end = Offset.zero;
-                        var curve = Curves.ease;
-
-                        var tween = Tween(begin: begin, end: end)
-                            .chain(CurveTween(curve: curve));
-
-                        return SlideTransition(
-                          position: animation.drive(tween),
-                          child: child,
-                        );
-                      },
-                    ),
-                  );
+                  navigateHorizontally(context: context, widget: nextPage!);
                 } else if (func != null && nextPage == null) {
                   func!();
                 }
